+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Change Log
 
+### 0.18.5: August 3, 2019
+- Bug : Fixed issue on Weight fragment adding unexpected value when changing date
+- Enhancement : Added graph in weight fragment
+- Bug : Fixed issue not showing the keyboard on WOrkout fragment (#86)
+- Enhancement : Added seconds to Duration for cardio records (#83)
 
 ### 0.18.4: July 22, 2019
 - Enhancement : Removed Google play services for FDroid compatibility (#84)